feat: Improve react router
Due that the previous code was displaying a white screen with nothing
displayed on it and according to udacity's mentor orientation, it was
looked for React Router documents at
reacttraing.com/react-router/web/guides/quick-start, trying to improve
the code concerning to React Router. In a first approach, it was
successful to display some of what the project asks for, but still is
not ideal.